除算の小数点の不具合を修正

+++ Calculator/Calc.cs
@@ -41,7 +41,7 @@ namespace Calculator
             }
 
             double z;
-            z = x / y;
+            z = (double)x / y;
             return z;
         }
